CHE-781:Move all platform-api-client-gwt to the che-core-ide-api module
