+++ assembly/assembly-ide-war/src/main/resources/org/eclipse/che/ide/IDE.gwt.xml
@@ -67,15 +67,15 @@
     <inherits name='org.eclipse.che.plugin.artik.ide.info.ProductInfo'/>
 
     <!-- Platform API GWT client dependencies                       -->
-    <inherits name='org.eclipse.che.api.Auth'/>
-    <inherits name='org.eclipse.che.api.Core'/>
-    <inherits name='org.eclipse.che.api.Machine'/>
-    <inherits name='org.eclipse.che.api.Git'/>
-    <inherits name='org.eclipse.che.api.Project'/>
-    <inherits name='org.eclipse.che.api.User'/>
-    <inherits name='org.eclipse.che.api.Factory'/>
-    <inherits name="org.eclipse.che.api.Workspace"/>
-    <inherits name="org.eclipse.che.api.Ssh"/>
+    <inherits name='org.eclipse.che.api.auth.Auth'/>
+    <inherits name='org.eclipse.che.api.core.Core'/>
+    <inherits name='org.eclipse.che.api.machine.Machine'/>
+    <inherits name='org.eclipse.che.api.git.Git'/>
+    <inherits name='org.eclipse.che.api.project.Project'/>
+    <inherits name='org.eclipse.che.api.user.User'/>
+    <inherits name='org.eclipse.che.api.factory.Factory'/>
+    <inherits name="org.eclipse.che.api.workspace.Workspace"/>
+    <inherits name="org.eclipse.che.api.ssh.Ssh"/>
 
     <stylesheet src="IDE.css"/>
