+++ plugins/plugin-artik-ide/src/main/java/org/eclipse/che/plugin/artik/ide/discovery/DeviceDiscoveryServiceClientImpl.java
@@ -13,9 +13,9 @@ package org.eclipse.che.plugin.artik.ide.discovery;
 
 import com.google.inject.Inject;
 
-import org.eclipse.che.api.machine.gwt.client.DevMachine;
 import org.eclipse.che.api.promises.client.Promise;
 import org.eclipse.che.ide.api.app.AppContext;
+import org.eclipse.che.ide.api.machine.DevMachine;
 import org.eclipse.che.ide.rest.AsyncRequestFactory;
 import org.eclipse.che.ide.rest.DtoUnmarshallerFactory;
 import org.eclipse.che.ide.ui.loaders.request.LoaderFactory;
