+++ plugins/plugin-artik-ide/src/main/resources/org/eclipse/che/plugin/artik/Artik.gwt.xml
@@ -21,7 +21,7 @@
     <inherits name='org.eclipse.che.ide.Api'/>
     <inherits name='org.eclipse.che.ide.ui.CodenvyUI'/>
     <inherits name="com.google.gwt.inject.Inject"/>
-    <inherits name="org.eclipse.che.api.Project"/>
+    <inherits name="org.eclipse.che.api.project.Project"/>
     <inherits name="org.eclipse.che.ide.Core"/>
 
     <source path="ide"/>
